Don't support strongly-typed parse results
Unfortunately, getting real mileage out of strongly-typed parse results
is hard.  This is becuase, first and foremost, Swift reflection isn't
readwrite (📡22209264).  You could get rewrite reflection via ObjC, but
the Linux story isn't well-established.

Strongly-typed stuff only has a benefit if it actually reduces the
number of opportunities to make typos (otherwise it's no better than
stringly-typed).  In practice, the attempted implementation does not.

However unlike, let’s say, a “real” Model, the parse results “built to
throw away”—e.g., it is used in one place, the place that receives this
parse, and not anywhere else

Now previously I would do you one of these

```swift
        private mutating func setValue(value: Any?, forKey key: String) {
            switch(key) {
            case "identityFile":
                identityFile = value as! String
            case "fisaFile":
                identityFile = value as! String
            case "identityDescription":
                identityDescription = value as! String
            default:
                preconditionFailure("Unknown key \(key)")
            }
        }
        private static func typeForKey(key: String) -> Any.Type {
            switch(key) {
            case "identityFile", "fisaFile", "identityDescription":
                return String.self
            default:
                preconditionFailure("Unknown key \(key)")
            }
        }
```

The thing is, I don’t think I’m gaining any mileage out of this pattern
in the “built to throw away” scenario.  If I just work with (let’s say)
a dictionary, there are exactly 2 places to misspell “identityFile”—when
I put it in and when I take it out of the dictionary.  And in the
`setValue` pattern there are also two places—inside `setValue` and
inside `typeForKey`.  So I don’t think I’ve bought myself anything (in
the throw-away scenario) with this approach.

As a result, I'm changing the implementation to be a ParseResult that is
essentially a dictionary of OptionTypes--stringly typed via the
subscript operator.  In practice, that eliminates a lot of code that
used to be here, and is probably a net improvement to the overall
library.