Remove a few optionals from our API.
In practice, when these optionals return nil, that's (application)
programmer error. You typoed a string name in your program somewhere.
Crashing is an appropriate behavior for programmer error.

An optional is not a good API choice here because asking programs to
sprinkle ! everywhere is ugly, asking for trouble, and asking
application programmers to make particular decisions about whether or
not they should handle the nil case when we already know they should
not.  Programmers should not be handling these conditions, they should
not be mistyping key strings in the first place.

Similarly an implicitly unwrapped optional was rejected, since A)
existing programs should remove their bangs (implicitly unwrapped
optionals don't introduce compile errors into existing programs, making
this hard to spot.)  And B) programs should not be writing code to
handle if these values are nil, they should be spelling keys correctly.
So we don't give them an API to even find out if they're nil, we just
crash.

Some day we will get proper Swift reflection (📡22209264) and maybe this
whole decision can be simplified and we can do away with stringly-typed
keys.  But until that day, let's do the best we can, hard-crashing if
the application programmer screws up writing their code, and not making
them do muscle-memory `if let` in a case that they should not handle,
they should just fix.