Update documentation
Writing documentation for most of the major CLIKit APIs.  Let's do a
brain dump while it's fresh in our heads.

Updating license headers in certain places.

The CommandParser is no longer a public API, I don't think it is
seriously used now that we have EasyCommand and I want to get this right
before we go to 1.0